requirements.txt
humanfriendly==10.0
psutil==7.0.0
python-dotenv==1.1.0
speedtest-cli==2.1.3
uptime==3.0.1
